Minor test refactoring.
